move op execution functions to executor.rs, delete block_handler.rs
lint
